added failure cases for definion ops

+++ haloSoftware.py
@@ -39,23 +39,45 @@ def login(username, password):
         if pw == password:
             user = username
             success = True
+            
+def logout():
+    global loggedIn
+    global success
+    global userlog
+    global user
+    if not loggedIn:
+        success = False
+    else:
+        userlog=user
+        user = None  
+        loggedIn = False
+        success = True
     
+def createEmptyFile(filename):
+    with open(filename, 'w') as f:
+        for i in range(1,9):
+            f.write("1,"+str(i)+",$,$,$,$,$,$,$,$,$,$,$\n")
+            for j in range(11):
+                if i==8 and j==10:
+                    f.write(",,,,,,,,,,,")
+                else:
+                    f.write(",,,,,,,,,,,\n")
                 
 def createNewFile(typename):
-    df = pd.read_csv("SystemCatalog.csv", index_col="type") #get type info
+    df = pd.read_csv("SystemCatalogue.csv", index_col="type") #get type info
     filenum = df.loc[typename].get("filenum")
     files = df.loc[typename].get("files")
     
     #create file
     filename = typename+str(filenum)+".txt"
-    f = open(filename,'w')
+    createEmptyFile(filename)
     
     df.loc[typename, "filenum"] =filenum + 1                #update system catalog
     df.loc[typename, "files"] = files + " " + filename
-    df.to_csv("SystemCatalog.csv")
+    df.to_csv("SystemCatalogue.csv")
     
 def deleteLastFile(typename):
-    df = pd.read_csv("SystemCatalog.csv", index_col="type") #get file names
+    df = pd.read_csv("SystemCatalogue.csv", index_col="type") #get file names
     files = df.loc[typename].get("files")
     filelist = files.split()
     filenum = df.loc[typename].get("filenum")
@@ -65,12 +87,12 @@ def deleteLastFile(typename):
     
     df.loc[typename, "filenum"] =filenum - 1                #update system catalog
     df.loc[typename, "files"] = " ".join(filelist[:-1])
-    df.to_csv("SystemCatalog.csv")
+    df.to_csv("SystemCatalogue.csv")
     
     
 def createSystemFiles():
     col = pd.DataFrame(columns=["type","filenum","files","fieldnum","fields"])
-    col.to_csv("SystemCatalog.csv", index=False)
+    col.to_csv("SystemCatalogue.csv", index=False)
     
     col = pd.DataFrame(columns=["username","occurrence","operation","status"])
     col.to_csv("haloLog.csv", index=False)
@@ -79,36 +101,59 @@ def createSystemFiles():
     col.to_csv("users.csv", index=False)
    
 def createType(operation):
+    global success
     typename = operation[0]
-    fieldnum = int(operation[1])+2
-    fields = ["planet", "id"]+operation[2:]
-    newType = {"type":[typename],"filenum":[1],"files":[typename+"0.txt"],"fieldnum":[fieldnum],"fields":[" ".join(fields)]}
-    data = pd.DataFrame(data=newType)
-    data.to_csv("SystemCatalog.csv", mode='a', index=False, header=False)
-    
-    filename = typename+"0.txt"
-    f = open(filename,'w')
+    df = pd.read_csv("SystemCatalogue.csv")
+    types = df["type"].tolist()
+    if typename in types:
+        success = False
+    else:
+        fieldnum = int(operation[1])+2
+        fields = ["planet", "id"]+operation[2:]
+        newType = {"type":[typename],"filenum":[1],"files":[typename+"0.txt"],"fieldnum":[fieldnum],"fields":[" ".join(fields)]}
+        data = pd.DataFrame(data=newType)
+        data.to_csv("SystemCatalogue.csv", mode='a', index=False, header=False)
+        df = pd.read_csv("SystemCatalogue.csv", index_col="type")    #sort
+        df.sort_index(inplace=True)
+        df.to_csv("SystemCatalogue.csv")
+        filename = typename+"0.txt"
+        createEmptyFile(filename)
+        success = True
     
 def inheritType(operation):
+    global success
     newType = operation[0]
     oldType = operation[1]
-    newFields = operation[2:]
-    df = pd.read_csv("SystemCatalog.csv", index_col="type")
-    fields = df.loc[oldType].get("fields").split()[2:]
-    fieldnum = df.loc[oldType].get("fieldnum") + len(newFields)
-    createType([newType]+[fieldnum]+fields+newFields)
+    df = pd.read_csv("SystemCatalogue.csv")
+    types = df["type"].tolist()
+    if newType in types or oldType not in types:
+        success = False
+    else:
+        newFields = operation[2:]
+        df = pd.read_csv("SystemCatalogue.csv", index_col="type")
+        fields = df.loc[oldType].get("fields").split()[2:]
+        fieldnum = df.loc[oldType].get("fieldnum")-2 + len(newFields)
+        createType([newType]+[fieldnum]+fields+newFields)
+        success = True
     
 def deleteType(typeName):
-    df = pd.read_csv("SystemCatalog.csv", index_col="type")
-    files = df.loc[typeName].get("files").split()
-    for file in files:
-        os.remove(file)
-    df.drop(index=typeName, inplace=True)
-    df.to_csv("SystemCatalog.csv")
+    global success
+    df = pd.read_csv("SystemCatalogue.csv")
+    types = df["type"].tolist()
+    if typeName not in types:
+        success = False
+    else:
+        df = pd.read_csv("SystemCatalogue.csv", index_col="type")
+        files = df.loc[typeName].get("files").split()
+        for file in files:
+            os.remove(file)
+        df.drop(index=typeName, inplace=True)
+        df.to_csv("SystemCatalogue.csv")
+        success = True
     
 def listType():
     global success
-    df = pd.read_csv("SystemCatalog.csv")
+    df = pd.read_csv("SystemCatalogue.csv")
     types = df["type"].tolist()
     with open(outputFile, "w") as f:
         for name in types:
@@ -133,10 +178,8 @@ with open(inputFile, 'r') as f:         #read input file
             if success:
                 loggedIn = True
         elif operation[0]=="logout":
-            userlog=user
-            user = None  
-            loggedIn = False
-        elif not loggedIn:		#if not logged in pass operation
+            logout()
+        elif not loggedIn:		#if not logged in skip
              success = False
         else:
             if operation[1]=="type":                                    #HALO Definition Language Operations
